Separate out the tests of RouteErrorHandler
