+++ backend/src/app.test.ts
@@ -1,13 +1,8 @@
 import { describe, it, expect, beforeEach, beforeAll } from '@jest/globals'
 import request from 'supertest'
 import { ExpressAppInitializer } from './app'
-import {
-  Express,
-  Request as ExpressRequest,
-  Response as ExpressResponse,
-} from 'express'
+import { Express } from 'express'
 import { RouteService } from './route/route'
-import { RouteErrorHandler } from './route/util'
 import { assertErrorResponse } from './test_utils/assert'
 
 interface Duty {
@@ -227,94 +222,5 @@ describe('Async error handling', () => {
       const response = await request(app).patch('/error')
       expect(response.status).toBe(500)
     })
-  })
-})
-
-describe('RouteErrorHandler', () => {
-  class CustomError extends Error {}
-  class UnexpectedError extends Error {}
-
-  function addRootGetRoute(
-    handler: (req: ExpressRequest, res: ExpressResponse) => Promise<void>,
-  ) {
-    const routeService = new (class extends RouteService {
-      constructor() {
-        super()
-        this.get('/', handler)
-      }
-    })()
-
-    return ExpressAppInitializer.createNullApp({
-      extraRouteConfigs: [
-        {
-          path: '/',
-          routeService,
-        },
-      ],
-    })
-  }
-
-  it('should return 500 for unexpected error', async () => {
-    const app = addRootGetRoute(async (req, res) => {
-      try {
-        throw new UnexpectedError('Unexpected error')
-      } catch (err) {
-        return new RouteErrorHandler([
-          {
-            statusCode: 401,
-            typeOfError: CustomError,
-          },
-        ]).handle(err, res)
-      }
-    })
-    const response = await request(app).get('/')
-    expect(response.status).toBe(500)
-  })
-
-  it('should return templated response for expected error', async () => {
-    const app = addRootGetRoute(async (req, res) => {
-      try {
-        throw new CustomError('Default message')
-      } catch (err) {
-        return new RouteErrorHandler([
-          {
-            statusCode: 403,
-            typeOfError: CustomError,
-          },
-        ]).handle(err, res)
-      }
-    })
-    const response = await request(app).get('/')
-    assertErrorResponse(
-      {
-        status: 403,
-        message: 'Default message',
-      },
-      response,
-    )
-  })
-
-  it('should custom response contain custom message if template provided it', async () => {
-    const app = addRootGetRoute(async (req, res) => {
-      try {
-        throw new CustomError('Default message')
-      } catch (err) {
-        return new RouteErrorHandler([
-          {
-            statusCode: 402,
-            typeOfError: CustomError,
-            customMessage: 'Custom message',
-          },
-        ]).handle(err, res)
-      }
-    })
-    const response = await request(app).get('/')
-    assertErrorResponse(
-      {
-        status: 402,
-        message: 'Custom message',
-      },
-      response,
-    )
   })
 })
